[CHANGE] Made changes requested by raphael
Added doxy comments;
Renamed some files;
changed ivec3 logging;
