+++ legion/engine/physics/broadphasecollisionalgorithms/broadphasebruteforce.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <physics/broadphase_collision_algorithms/broadphasecollisionalgorithm.hpp>
+#include <physics/broadphasecollisionalgorithms/broadphasecollisionalgorithm.hpp>
 
 namespace legion::physics
 {
@@ -7,6 +7,10 @@ namespace legion::physics
     {
     public:
 
+        /**@brief Collects collider pairs that have a chance of colliding and should be checked in narrow-phase collision detection
+         * @param manifoldPrecursors all the physics components
+         * @param manifoldPrecursorGrouping a list-list of colliders that have a chance of colliding and should be checked
+         */
         void collectPairs(std::vector<physics_manifold_precursor>& manifoldPrecursors,
             std::vector<std::vector<physics_manifold_precursor>>& manifoldPrecursorGrouping) override
         {
